requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
pandas>=2.0.0
pathlib2>=2.3.7
tqdm>=4.65.0